Running a Streamlit server with Field Extraction
This file contains the code for running a Streamlit webapp on the Streamlit community server after going through a tunnel. It is fully functional and runs the YOLO v8 nano model in the backend to predict cheque fields and display them. This only has the problem that it crashes very often since the community on Streamlit is very active.